+++ app/Http/Controllers/Backend/Shop/ShopController.php
@@ -10,6 +10,7 @@ use App\Models\Shop\ShopType;
 use Repository\Shop\ShopRepository;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Repository\Location\AreaRepository;
 use Repository\Location\CityRepository;
 use Repository\Shop\ShopTypeRepository;
@@ -44,22 +45,16 @@ class ShopController extends Controller
         $this->shopMediaRepo= $shopMediaRepository;
     }
 
-    /**
-     * Display a listing of the resource.
-     * @return Renderable
-     */
     public function index()
     {
+        Gate::authorize('backend.shops.index');
         $shops = $this->shopRepo->getAll();
         return view('backend.shop.shop.index',  compact('shops'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     * @return Renderable
-     */
     public function create()
     {
+        Gate::authorize('backend.shops.create');
         $cities = $this->cityRepo->getAll();
         $areas = $this->areaRepo->getAll();
         $markets = $this->marketRepo->getAll();
@@ -67,11 +62,6 @@ class ShopController extends Controller
         return view('backend.shop.shop.form', compact('cities','areas' ,'markets', 'shoptypes'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     * @param Request $request
-     * @return Renderable
-     */
     public function store(Request $request)
     {
         $request->validate([
@@ -99,23 +89,14 @@ class ShopController extends Controller
         return redirect()->route('backend.shops.index');
     }
 
-    /**
-     * Show the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
     public function show($id)
     {
         return view('show');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
     public function edit($id)
     {
+        Gate::authorize('backend.shops.edit');
         $cities     = $this->cityRepo->getAll();
         $areas      = $this->areaRepo->getAll();
         $markets    = $this->marketRepo->getAll();
@@ -124,23 +105,15 @@ class ShopController extends Controller
         return view('backend.shop.shop.form', compact('cities','areas', 'markets', 'shoptypes', 'shop'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     * @param Request $request
-     * @param int $id
-     * @return Renderable
-     */
     public function update(Request $request, $id)
     {
         $shop = $this->shopRepo->updateByShopOwner($id);
-
         if ($shop == null) {
             return $this->bad(
                 'You are unauthorized to update this shop!',
                 403
             );
         }
-
         $shopLogo       = $request->hasFile('logo');
         $shopCoverImage = $request->hasFile('cover_image');
         $metaImageShop  = $request->hasFile('meta_og_image');
@@ -152,15 +125,12 @@ class ShopController extends Controller
         if ($shopLogo) {
             $this->shopRepo->updateShopsLogo($id);
         }
-
         if ($shopCoverImage) {
             $this->shopRepo->updateShopsImage($id);
         }
-
         if ($metaImageShop) {
             $this->shopRepo->updateShopsOgImage($id);
         }
-
         $this->shopRepo->updateByID($id, $request->except('logo', 'cover_image', 'meta_og_image') +
             [
                 'logo' => $logo,
@@ -168,18 +138,13 @@ class ShopController extends Controller
                 'meta_og_image' => $meta_og_image
             ]);
         $this->shopMediaRepo->shopGalleryUpdate($request->hasFile('image') ? $request->file('image') : $shop->shopMedia, $id);
-
         notify()->success('Shop Successfully Updated.', 'Updated');
         return redirect()->route('backend.shops.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     * @param int $id
-     * @return Renderable
-     */
     public function destroy($id)
     {
+        Gate::authorize('backend.shops.destroy');
         $this->shopRepo->deleteShops($id);
         notify()->success('Shop Successfully Deleted.', 'Deleted');
         return redirect()->route('backend.shops.index');
