+++ app/Http/Controllers/Backend/Shop/ShopTypeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend\Shop;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Repository\Shop\ShopTypeRepository;
 
 class ShopTypeController extends Controller
@@ -15,30 +16,19 @@ class ShopTypeController extends Controller
         $this->shopTypeRepo = $shopTypeRepository;
     }
 
-    /**
-     * Display a listing of the resource.
-     * @return Renderable
-     */
     public function index()
     {
+        Gate::authorize('backend.shoptypes.index');
         $shoptypes  = $this->shopTypeRepo->getAll();
         return view('backend.shop.shoptype.index',  compact('shoptypes'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     * @return Renderable
-     */
     public function create()
     {
+        Gate::authorize('backend.shoptypes.create');
         return view('backend.shop.shoptype.form');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     * @param Request $request
-     * @return Renderable
-     */
     public function store(Request $request)
     {
         $request->validate([
@@ -46,53 +36,32 @@ class ShopTypeController extends Controller
         ]);
 
         $this->shopTypeRepo->create($request->all());
-
         notify()->success('Shop type Successfully Added.', 'Added');
         return redirect()->route('backend.shoptypes.index');
     }
 
-    /**
-     * Show the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
     public function show($id)
     {
         return view('show');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
     public function edit($id)
     {
+        Gate::authorize('backend.shoptypes.edit');
         $shoptype = $this->shopTypeRepo->findByID($id);
         return view('backend.shop.shoptype.form', compact('shoptype'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     * @param Request $request
-     * @param int $id
-     * @return Renderable
-     */
     public function update(Request $request, $id)
     {
         $this->shopTypeRepo->updateByID($id, $request->all());
-
         notify()->success('Shop type Successfully Updated.', 'Updated');
         return redirect()->route('backend.shoptypes.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     * @param int $id
-     * @return Renderable
-     */
     public function destroy($id)
     {
+        Gate::authorize('backend.shoptypes.destroy');
         $this->shopTypeRepo->deletedByID($id);
         notify()->warning('Floor Successfully Deleted.', 'Deleted');
         return redirect()->route('backend.shoptypes.index');
